+++ .gitignore
@@ -3,3 +3,5 @@ pdf.pdf
 intelisa.pdf
 openweb_tm-PRINT.pdf
 local.mk
+build/
+
