+++ web/viewer.html
@@ -6,25 +6,7 @@
 
         <script type="text/javascript" src="compatibility.js"></script>
         <script type="text/javascript" src="viewer.js"></script>
-        <script type="text/javascript" src="../src/pdf.js"></script>
-        <script type="text/javascript" src="../src/util.js"></script>
-        <script type="text/javascript" src="../src/canvas.js"></script>
-        <script type="text/javascript" src="../src/function.js"></script>
-        <script type="text/javascript" src="../src/colorspace.js"></script>
-        <script type="text/javascript" src="../src/image.js"></script>
-        <script type="text/javascript" src="../src/pattern.js"></script>
-        <script type="text/javascript" src="../src/obj.js"></script>
-        <script type="text/javascript" src="../src/parser.js"></script>
-        <script type="text/javascript" src="../src/evaluator.js"></script>
-        <script type="text/javascript" src="../src/stream.js"></script>
-        <script type="text/javascript" src="../src/fonts.js"></script>
-        <script type="text/javascript" src="../src/crypto.js"></script>
-        <script type="text/javascript" src="../src/glyphlist.js"></script>
-        <script type="text/javascript" src="../src/metrics.js"></script>
-        <script type="text/javascript" src="../src/charsets.js"></script>
-        <script type="text/javascript" src="../src/cidmaps.js"></script>
-        <script type="text/javascript" src="../src/worker/message_handler.js"></script>
-        <script type="text/javascript" src="../src/worker/processor_handler.js"></script>
+        <script type="text/javascript" src="../pdf.js"></script>
   </head>
 
   <body>
